% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{em}
\alias{em}
\title{EM algorithm to compute the MLE for beta regression with rounded responses}
\usage{
em(y, x, theta, maxit = 30L, eps = 1e-07, verbose = TRUE)
}
\arguments{
\item{y}{a n-vector of response}

\item{x}{a n x p matrix of design}

\item{theta}{vector of parameter}

\item{maxit}{maximum number of iteration}

\item{eps}{tolerance}

\item{verbose}{boolean for printing information}
}
\description{
EM algorithm to compute the MLE for beta regression with rounded responses
}
