% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_negbin.R
\name{glm_negbin}
\alias{glm_negbin}
\title{Function to estimate consistent MLE for negative binomial}
\usage{
glm_negbin(y, x, lambda, maxit = 50L, epsilon = 1e-07, trace = FALSE)
}
\arguments{
\item{y}{response vector}

\item{x}{design matrix}

\item{lambda}{mean parameter for Poisson censoring process}

\item{maxit}{maximum number of iteration}

\item{epsilon}{tolerance parameter}

\item{trace}{boolean for printing information}
}
\description{
Function to estimate consistent MLE for negative binomial
}
