% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inverse_FIM}
\alias{inverse_FIM}
\title{Inverse Fisher information matrix for logistic regression with misclassified
responses}
\usage{
inverse_FIM(x, beta, fp, fn)
}
\arguments{
\item{x}{a n x p matrix of design}

\item{beta}{a p-vector of parameter}

\item{fp}{false positive rate}

\item{fn}{false negative rate}
}
\description{
Inverse Fisher information matrix for logistic regression with misclassified
responses
}
