% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_misclassification_mle}
\alias{logistic_misclassification_mle}
\title{MLE for logistic regression with misclassified responses}
\usage{
logistic_misclassification_mle(x, y, fp, fn)
}
\arguments{
\item{x}{a n x p matrix of design}

\item{y}{a n-vector of response}

\item{fp}{false positive rate}

\item{fn}{false negative rate}
}
\description{
MLE for logistic regression with misclassified responses
}
