% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make_negbin}
\alias{make_negbin}
\title{Construct a `negbin` object for negative binomial regression with predefined coefficients}
\usage{
make_negbin(x, coefficients, theta)
}
\arguments{
\item{x}{n x p design matrix (no intercept)}

\item{coefficients}{p+1 vector of coefficients, the first coefficient is the intercept.}

\item{theta}{additional parameter for negative binomial regression}
}
\description{
It creates an object of class \code{negbin} from \code{MASS} package,
as one would obtain when fitting a negative binomial regression, but here there is no
fit, one simply provides a design matrix and coefficients. This function is
particularly useful in conjunction with \code{simulation} method from the
\code{ib} package to simulate from this model.
}
\seealso{
\code{\link[MASS]{glm.nb}}, \code{\link[ib]{simulation}}
}
