% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nll_max_alpha}
\alias{nll_max_alpha}
\title{Negative log-likelihood for the overdispersion parameter (negative binomial regression)}
\usage{
nll_max_alpha(alpha, beta, y, x, lambda)
}
\arguments{
\item{alpha}{parameter of negative binomial}

\item{beta}{a p-vector of coefficients}

\item{y}{a n-vector of response}

\item{x}{a n x p matrix of design}

\item{lambda}{mean of Poisson censoring process}
}
\description{
Negative log-likelihood for the overdispersion parameter (negative binomial regression)
}
