% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nll_max_beta}
\alias{nll_max_beta}
\title{negative log-likelihood for beta coefficients}
\usage{
nll_max_beta(beta, sigma, y, x, lambda)
}
\arguments{
\item{beta}{a p-vector of coefficients}

\item{sigma}{parameter of negative binomial}

\item{y}{a n-vector of response}

\item{x}{a n x p matrix of design}

\item{lambda}{mean of Poisson censoring process}
}
\description{
negative log-likelihood for beta coefficients
}
