% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nll_max_sigma}
\alias{nll_max_sigma}
\title{negative log-likelihood for sigma}
\usage{
nll_max_sigma(sigma, beta, y, x, lambda)
}
\arguments{
\item{sigma}{parameter of negative binomial}

\item{beta}{a p-vector of coefficients}

\item{y}{a n-vector of response}

\item{x}{a n x p matrix of design}

\item{lambda}{mean of Poisson censoring process}
}
\description{
negative log-likelihood for sigma
}
