% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{par_bootstrap_mle}
\alias{par_bootstrap_mle}
\title{Parametric bootstrap for logistic regression with misclassified responses}
\usage{
par_bootstrap_mle(beta, x, B, seed, ncores, fp = 0, fn = 0)
}
\arguments{
\item{beta}{a p-vector of parameter}

\item{x}{a n x p design matrix}

\item{B}{the number of bootstrap replicates}

\item{seed}{for random number generator}

\item{ncores}{number of cores for parallelisation}

\item{fp}{false positive rate}

\item{fn}{false negative rate}
}
\description{
Parametric bootstrap for logistic regression with misclassified responses
}
