% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_negbin.R
\name{glm_negbin}
\alias{glm_negbin}
\title{MLE for negative binomial with interfered responses}
\usage{
glm_negbin(y, x, lambda, maxit = 50L, epsilon = 1e-07, trace = FALSE)
}
\arguments{
\item{y}{response vector}

\item{x}{design matrix}

\item{lambda}{mean parameter for Poisson censoring process}

\item{maxit}{maximum number of iteration}

\item{epsilon}{tolerance parameter}

\item{trace}{boolean for printing information}
}
\description{
Implementation of the MLE for negative binomial regression
with interfered responses: one observes \eqn{\max(y_i,z_i)},
where \eqn{y_i} is the negative binomial response and \eqn{z_i}
is a Poisson random variable with mean \eqn{\lambda}.
}
